add a new problem
namespace tests;

public class BestTimeToBuySellStockTests
{

    [Fact]
    public void CalculateProfit_ReturnsFiveForScenario1()
    {
        int[] prices = [7, 1, 5, 3, 6, 4];
        BestTimeToBuySellStock stock = new();
        const int expected = 5;

        int actual = stock.CalculateMaximumProfit(prices);

        actual.Should().Be(expected);
    }
}

/*
You are given an array of prices
where prices[i] is the price of a given stock on the ith day.
You want to maximize your profit
by choosing a single day
to buy one stock
and choosing a different day in the future
to sell that stock.

Input: prices = [7,1,5,3,6,4];
Output: 5
Explanation: Buy on day 2 (price = 1) and sell on day 5 (price = 6), profit = 6 - 1 = 5.
Note that buying on day 2 and selling on day 1 is not allowed because you must buy before you sell.
*/
namespace hellolib;

public class BestTimeToBuySellStock
{
    public int CalculateMaximumProfit(int[] prices)
    {
        int minPrice = int.MaxValue;
        int maxProfit = 0;
        foreach (int price in prices)
        {
            if (price < minPrice)
            {
                minPrice = price;
            }
            int profit = price - minPrice;
            if (profit > maxProfit)
            {
                maxProfit = profit;
            }
        }
        return maxProfit;
    }
}